feat(kb_basic_progression): point to next steps
Rather than continue to build up the complexity of the basic
progression, the plan is to split off into a second diagram that covers
the TGU skills that are developed from KB-24 on. I'm including four of
these with different colourings so that it's easy to track from one
diagram to another.

Also:
- Note what the start stop drills achieve and why they are important
  for subsequent drills