refactor: create src structure
Having every diagram in root was starting to become annoying, so I've
put it all into a `src` directory structure. This necessitated updating
the Readme with a nice new table format for each diagram. Makes it a
little bit easier for someone to grab what they want, I hope.

This also becomes a precursor to a simple build process.