build(psakefile): enable setting output format
Instead of generating hardcoded SVGs, let's allow our build to pass in
an outputFormat and pass that on to Dot and PlantUML. Because PlantUML
actually uses `-tsvg and `tpng` as flags rather than string modifiers, I
needed to switch to using Invoke-Expression. And that was so hefty that
I decided to make a function, which I then stored in a separate file.

Also, `Clean` now actually deletes all the files in the dist directory,
so be aware of this change lol.