feat(2hhc_bootstrap_program): initial commit
Add a two-handed heavy club program based on progression of skills
rather than increasing volume or intensity. The idea is for someone who
has never encountered heavy club before to be able to watch some videos
and start swinging and building foundational competence before switching
to a strength and conditioning program.

I'm using XLSX format reluctantly as I want people to be able to open
the damn thing from Excel and have it be basically legible.