feat(kb_bootstrap_program): initial 9 sessions
Take the idea of bootstrapping a program from the 2HHC section and apply
it to getting started with kettlebells. The approach is a little
different as MW has big subsequences dedicated to e.g. the C&P, but we
can do a bit of Tetris of Training to alternate swings and clean days etc