feat(mace_warrior_subsequence): add Sozi
Sozi allows one to build up animated presentations that traverse an SVG
image, and so this experiment demonstrates a quick flyover of the
warrior stance sub-sequence to show how it works. I *think* just the
JSON is required to be saved, with the two HTML files generated from it,
but saving all three to be safe.