feat(LICENSE): add project-wide license
Although I already had licenses on the diagrams, adding a copy of the
CC BY-NC-SA 4.0 legal terms feels like a good and important thing to do.

I guess the main thing this changes is it releases the non-diagram work
under the same license, which was the original intention in any case.