feat(hc_2h_progression): rm _2
The second breakout of the two-handed heavy club sequence is not a great
diagram and it doesn't serve much use, so I'm removing it. I'm keeping
the first breakout as there is a lot of value in the bootstrap/first
steps diagram alongside the _ALL view.

Update build and Readme to no longer reference the file.